# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Distill API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/needs": {
      "get": {
        "tags": [
          "needs"
        ],
        "summary": "Find needs",
        "operationId": "list_needs_needs_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "Text in the need or its member requests",
              "title": "Q"
            },
            "description": "Text in the need or its member requests"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/NeedStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "product_area",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Area"
            }
          },
          {
            "name": "segment",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Segment"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Needs with a request or confirmed support from it",
              "title": "Segment"
            },
            "description": "Needs with a request or confirmed support from it"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "priority",
                "support",
                "recent"
              ],
              "type": "string",
              "default": "priority",
              "title": "Sort"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeedPage"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/needs/similar": {
      "get": {
        "tags": [
          "needs"
        ],
        "summary": "Needs similar to what the requester is typing (embeddings only)",
        "operationId": "similar_needs_needs_similar_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 500,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimilarNeed"
                  },
                  "title": "Response Similar Needs Needs Similar Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/needs/{need_id}": {
      "get": {
        "tags": [
          "needs"
        ],
        "summary": "Get Need",
        "operationId": "get_need_needs__need_id__get",
        "parameters": [
          {
            "name": "need_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Need Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeedDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/needs/{need_id}/support": {
      "post": {
        "tags": [
          "needs"
        ],
        "summary": "Add support (a requester claim)",
        "description": "201 the first time, 200 on a repeat by the same requester. Not counted until confirmed. 409 if the need was merged into another one.",
        "operationId": "add_support_needs__need_id__support_post",
        "parameters": [
          {
            "name": "need_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Need Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupportCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportOut"
                }
              }
            }
          },
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportOut"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/needs/{need_id}/status": {
      "patch": {
        "tags": [
          "needs"
        ],
        "summary": "Change a need's status (a human product decision) and queue the stakeholder drafts",
        "description": "Saved at once; the worker then drafts a personal update per supporter and a CS note per account. Nothing is sent until a PM approves each draft. 409 for a merged need or the same status.",
        "operationId": "change_status_needs__need_id__status_patch",
        "parameters": [
          {
            "name": "need_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Need Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusChangeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusChangeOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/needs/{need_id}/updates": {
      "get": {
        "tags": [
          "needs"
        ],
        "summary": "Status changes with their drafted, approved and discarded stakeholder messages (PM view)",
        "operationId": "need_updates_needs__need_id__updates_get",
        "parameters": [
          {
            "name": "need_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Need Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeedUpdates"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/needs/{need_id}/status-changes/{change_id}/redraft": {
      "post": {
        "tags": [
          "needs"
        ],
        "summary": "Draft the stakeholder messages again after drafting failed (spends one model call)",
        "operationId": "redraft_needs__need_id__status_changes__change_id__redraft_post",
        "parameters": [
          {
            "name": "change_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Change Id"
            }
          },
          {
            "name": "need_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Need Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Decision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusChangeOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Decision": {
        "properties": {
          "by": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "By"
          }
        },
        "type": "object",
        "required": [
          "by"
        ],
        "title": "Decision",
        "description": "Who decided. There is no auth yet (spec A3), so the PM names themselves."
      },
...
      "ErrorInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorInfo"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NeedDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "problem": {
            "type": "string",
            "title": "Problem"
          },
          "persona": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona"
          },
          "product_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Area"
          },
          "status": {
            "$ref": "#/components/schemas/NeedStatus"
          },
          "priority_score": {
            "type": "number",
            "title": "Priority Score",
            "description": "0-100, computed when read from current data and config"
          },
          "breakdown": {
            "$ref": "#/components/schemas/PriorityBreakdown"
          },
          "support_count": {
            "type": "integer",
            "title": "Support Count",
            "description": "Distinct requesters with a member request or a confirmed support. Staff filing for several customers count once; account_count shows the customers."
          },
          "claimed_support_count": {
            "type": "integer",
            "title": "Claimed Support Count",
            "description": "Supports still waiting for the intake workflow to confirm them"
          },
          "request_count": {
            "type": "integer",
            "title": "Request Count"
          },
          "account_count": {
            "type": "integer",
            "title": "Account Count"
          },
          "last_request_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Request At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "job_to_be_done": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job To Be Done"
          },
          "merged_into_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merged Into Id"
          },
          "requests": {
            "items": {
              "$ref": "#/components/schemas/NeedRequestOut"
            },
            "type": "array",
            "title": "Requests"
          },
          "supports": {
            "items": {
              "$ref": "#/components/schemas/SupportOut"
            },
            "type": "array",
            "title": "Supports"
          },
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/AccountOut"
            },
            "type": "array",
            "title": "Accounts"
          },
          "origin": {
            "$ref": "#/components/schemas/NeedOrigin"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "updates": {
            "items": {
              "$ref": "#/components/schemas/UpdateOut"
            },
            "type": "array",
            "title": "Updates",
            "description": "Approved updates sent to supporters (drafts are never shown)"
          },
          "audit_trail": {
            "items": {
              "$ref": "#/components/schemas/AuditEvent"
            },
            "type": "array",
            "title": "Audit Trail",
            "description": "Links, unlinks, status changes, decisions and AI runs, newest first"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "problem",
          "persona",
          "product_area",
          "status",
          "priority_score",
          "breakdown",
          "support_count",
          "claimed_support_count",
          "request_count",
          "account_count",
          "last_request_at",
          "created_at",
          "job_to_be_done",
          "merged_into_id",
          "requests",
          "supports",
          "accounts",
          "origin",
          "evidence",
          "updates",
          "audit_trail"
        ],
        "title": "NeedDetail"
      },
...
      "NeedPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NeedSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "NeedPage"
      },
...
      "NeedStatus": {
        "type": "string",
        "enum": [
          "open",
          "planned",
          "in_progress",
          "shipped",
          "declined",
          "merged"
        ],
        "title": "NeedStatus"
      },
...
      "NeedUpdates": {
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/StatusChangeOut"
            },
            "type": "array",
            "title": "Changes",
            "description": "Newest first, each with its drafts"
          }
        },
        "type": "object",
        "required": [
          "changes"
        ],
        "title": "NeedUpdates"
      },
...
      "Segment": {
        "type": "string",
        "enum": [
          "enterprise",
          "mid_market",
          "smb"
        ],
        "title": "Segment"
      },
...
      "SimilarNeed": {
        "properties": {
          "need_id": {
            "type": "integer",
            "title": "Need Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "problem": {
            "type": "string",
            "title": "Problem"
          },
          "persona": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona"
          },
          "product_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Area"
          },
          "status": {
            "$ref": "#/components/schemas/NeedStatus"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "need_id",
          "title",
          "problem",
          "persona",
          "product_area",
          "status",
          "score"
        ],
        "title": "SimilarNeed"
      },
      "StatusChangeIn": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "open",
              "planned",
              "in_progress",
              "shipped",
              "declined"
            ],
            "title": "Status",
            "description": "merged is set only by merging"
          },
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason"
          },
          "target_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date",
            "description": "Set only if the PM commits to a date"
          },
          "by": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "By"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "reason",
          "by"
        ],
        "title": "StatusChangeIn",
        "description": "A product decision (spec §7): who, what, why, and a date only if the PM commits to one."
      },
      "StatusChangeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "need_id": {
            "type": "integer",
            "title": "Need Id"
          },
          "from_status": {
            "$ref": "#/components/schemas/NeedStatus"
          },
          "to_status": {
            "$ref": "#/components/schemas/NeedStatus"
          },
          "by": {
            "type": "string",
            "title": "By"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "target_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "drafts_status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "pending",
                  "drafted",
                  "failed",
                  "superseded"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Drafts Status"
          },
          "drafts_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drafts Error"
          },
          "updates": {
            "items": {
              "$ref": "#/components/schemas/DraftOut"
            },
            "type": "array",
            "title": "Updates"
          }
        },
        "type": "object",
        "required": [
          "id",
          "need_id",
          "from_status",
          "to_status",
          "by",
          "reason",
          "target_date",
          "created_at",
          "drafts_status",
          "drafts_error",
          "updates"
        ],
        "title": "StatusChangeOut"
      },
...
      "SupportCreate": {
        "properties": {
          "requester_id": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 1.0,
            "title": "Requester Id"
          },
          "why_it_matters": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Why It Matters",
            "description": "Omit on a repeat to keep the stored reason"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requester_id",
          "severity"
        ],
        "title": "SupportCreate"
      },
...
      "SupportOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "need_id": {
            "type": "integer",
            "title": "Need Id"
          },
          "requester_id": {
            "type": "integer",
            "title": "Requester Id"
          },
          "requester_name": {
            "type": "string",
            "title": "Requester Name"
          },
          "account_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Name"
          },
          "why_it_matters": {
            "type": "string",
            "title": "Why It Matters"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "link_status": {
            "$ref": "#/components/schemas/SupportLinkStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "need_id",
          "requester_id",
          "requester_name",
          "account_name",
          "why_it_matters",
          "severity",
          "link_status",
          "created_at"
        ],
        "title": "SupportOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}